requirements.txt
twine==3.1.1
Django==3.0.7
djangorestframework==3.11.0
